Fix PouchDB custom builds
Commit 7db429c305358567de8e624223f095728aeca04a[1] introduces PouchDB
custom builds.
However, there were still a reference to the full pouchdb package
in the dev-server (as can be seen in the commit diff).

This commit uses pouchdb-core to register pouchdb-adapter-http and
pouchdb-load as described in the blog post introducting custom builds[2].

As this is my first contribution to this project, I take the opportunity
to thank you for all the hard work you do on open source projects.
You are, at least for me, one of the most inspiring web dev.
Keep going !

[1] https://github.com/nolanlawson/pokedex.org/commit/7db429c305358567de8e624223f095728aeca04a
[2] https://pouchdb.com/2016/06/06/introducing-pouchdb-custom-builds.html