Revert "Merge pull request #53 from djrosenbaum/master"
This reverts commit 31c4f71ea90a3bdd7376ca41a654325a695f221b, reversing
changes made to 64587710b01652267162ba17843ddf53def2e7d2.

I'm temporarily reverting this because in order for this commit to actually
work, we need a way to propagate changes to the Pokedex data to those
who already have it saved locally in IDB. In other words, it's hard
and I'm going to punt on it for now.